Tested Accelerators and Decelerators with the move to X idea.
Tested the Move to X idea we had for the speed boosters and removers, but 1 Unity crashed so it didn't save which is my bad and 2 I couldn't get them to work anyways so that is good to know. If anybody wants to try doing it instead, go right ahead, but I am going to try manually changing the speed in game.